fixing broken links
for Discord,
I updated it with https://anilist.co 's footer's discord link. 
I think it would be better to have a link like "https://anilist.co/discord" that redirects to the discord invite link.

for GraphQL extensions,
I searched with the previous one's name and it came as this one, 
which is the most popular in the chrome web store too.